Further Abstraction
- Refactored parts of the Command Framework to have much improved
abstraction.
- Command prefixes are now created procedurally instead of baked.
- Categories are now individual objects instead of enums, allowing for
customized commands instead of limited to the few standard ones. Old
standard categories has been reimplemented through this system.
- Commands "Execute" method parameters are now saved in an intermediate
"Parameter" class, which allows for better control of it, since it no
longer takes directly from the code layer. For example, it can be used
to overwrite which types an overload takes in, such as taking in
specific types instead of an params object[]
It is not completely done, but I've run out of time for the moment.